refactor(ts-sdk): fold position preview into order-preview
Review feedback: one preview module, not two — the liquidation/pro-forma
functions move into order-preview.ts unchanged, sharing the position
lookup and abs helpers. No behavior change.
